build: fix char-subscript and 64-to-32 narrowing warnings
- The MusicDetailView subclasses index levelNumImg with the char level fields
  (levelBas/Adv/Ext and the char level local); cast those subscripts to int so
  the index is unambiguous.
- The Downloader delegate callbacks read [downloader tag] through an id receiver;
  cast to int to match the Downloader.tag width the binary stores.
- Match the argument widths at three call sites: VerifyMd5Digest takes an
  unsigned int length; Texture2D -spriteAtIndex: takes an unsigned int, so the
  NSUInteger sprite index and the drawIndex cast narrow to unsigned int.

+++ Project/EditNoteRendererPhone.m
@@ -551,7 +551,7 @@ static inline void EditNoteRendererPhoneRenderFullcomboOut(
     drawPosition:(CGPoint)drawPosition
         drawArea:(CGRect)drawArea
            alpha:(float)alpha {
-    CGRect sprite = [self.texFront spriteAtIndex:(NSUInteger)drawIndex];
+    CGRect sprite = [self.texFront spriteAtIndex:(unsigned int)drawIndex];
     // The chip runs from drawPosition over the sprite's size; the visible portion is the
     // intersection with drawArea. If they do not overlap in either axis, nothing is drawn.
     double areaLeft = (double)(float)drawArea.origin.x;
